fix: replace leaflet::panTo() with setView()
This commit fixes the behaviour if the location is slightly off screen.
On clicking the track chart... the view is set to the clicked location
regardless of the current extent.
It also fixes a bug where wrong extent is being shown on single click
after zoom out by scrolling at the edge.

Signed-off-by: t0b3 <thomas.bettler@gmail.com>